feat: Support defaults for chat/completions request parameters in config #351
