+++ src/main/java/com/epam/aidial/core/controller/DeploymentPostController.java
@@ -2,9 +2,7 @@ package com.epam.aidial.core.controller;
 
 import com.epam.aidial.core.Proxy;
 import com.epam.aidial.core.ProxyContext;
-import com.epam.aidial.core.config.Addon;
 import com.epam.aidial.core.config.ApiKeyData;
-import com.epam.aidial.core.config.Assistant;
 import com.epam.aidial.core.config.Config;
 import com.epam.aidial.core.config.Deployment;
 import com.epam.aidial.core.config.Model;
@@ -12,8 +10,12 @@ import com.epam.aidial.core.config.ModelType;
 import com.epam.aidial.core.config.Pricing;
 import com.epam.aidial.core.config.Upstream;
 import com.epam.aidial.core.data.ErrorData;
+import com.epam.aidial.core.function.BaseFunction;
+import com.epam.aidial.core.function.CollectAttachmentsFn;
+import com.epam.aidial.core.function.enhancement.ApplyDefaultDeploymentSettingsFn;
+import com.epam.aidial.core.function.enhancement.EnhanceAssistantRequestFn;
+import com.epam.aidial.core.function.enhancement.EnhanceModelRequestFn;
 import com.epam.aidial.core.limiter.RateLimitResult;
-import com.epam.aidial.core.security.AccessService;
 import com.epam.aidial.core.storage.ResourceDescription;
 import com.epam.aidial.core.token.TokenUsage;
 import com.epam.aidial.core.token.TokenUsageParser;
@@ -21,12 +23,9 @@ import com.epam.aidial.core.upstream.DeploymentUpstreamProvider;
 import com.epam.aidial.core.upstream.UpstreamProvider;
 import com.epam.aidial.core.upstream.UpstreamRoute;
 import com.epam.aidial.core.util.BufferingReadStream;
-import com.epam.aidial.core.util.HttpException;
 import com.epam.aidial.core.util.HttpStatus;
 import com.epam.aidial.core.util.ModelCostCalculator;
 import com.epam.aidial.core.util.ProxyUtil;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.annotations.VisibleForTesting;
 import io.netty.buffer.ByteBufInputStream;
@@ -39,7 +38,6 @@ import io.vertx.core.http.HttpHeaders;
 import io.vertx.core.http.HttpServerRequest;
 import io.vertx.core.http.HttpServerResponse;
 import io.vertx.core.http.RequestOptions;
-import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
@@ -48,15 +46,12 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.math.BigDecimal;
 import java.net.URI;
-import java.util.HashMap;
-import java.util.LinkedHashSet;
-import java.util.Map;
+import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 import java.util.function.Function;
 
 @Slf4j
-@RequiredArgsConstructor
 public class DeploymentPostController {
 
     private static final Set<Integer> RETRIABLE_HTTP_CODES = Set.of(HttpStatus.TOO_MANY_REQUESTS.getCode(),
@@ -65,6 +60,17 @@ public class DeploymentPostController {
 
     private final Proxy proxy;
     private final ProxyContext context;
+
+    private final List<BaseFunction<ObjectNode>> enhancementFunctions;
+
+    public DeploymentPostController(Proxy proxy, ProxyContext context) {
+        this.proxy = proxy;
+        this.context = context;
+        this.enhancementFunctions = List.of(new CollectAttachmentsFn(proxy, context),
+                new ApplyDefaultDeploymentSettingsFn(proxy, context),
+                new EnhanceAssistantRequestFn(proxy, context),
+                new EnhanceModelRequestFn(proxy, context));
+    }
 
     public Future<?> handle(String deploymentId, String deploymentApi) {
         String contentType = context.getRequest().getHeader(HttpHeaders.CONTENT_TYPE);
@@ -205,50 +211,10 @@ public class DeploymentPostController {
 
         try (InputStream stream = new ByteBufInputStream(requestBody.getByteBuf())) {
             ObjectNode tree = (ObjectNode) ProxyUtil.MAPPER.readTree(stream);
-
-            try {
-                ProxyUtil.collectAttachedFiles(tree, this::processAttachedFile);
-                // assign api key data after processing attachments
-                ApiKeyData destApiKeyData = context.getProxyApiKeyData();
-                proxy.getApiKeyStore().assignPerRequestApiKey(destApiKeyData);
-            } catch (HttpException e) {
-                respond(e.getStatus(), e.getMessage());
-                log.warn("Can't collect attached files. Trace: {}. Span: {}. Error: {}",
-                        context.getTraceId(), context.getSpanId(), e.getMessage());
+            Throwable error = ProxyUtil.processChain(tree, enhancementFunctions);
+            if (error != null) {
+                finalizeRequest();
                 return;
-            } catch (Throwable e) {
-                context.respond(HttpStatus.BAD_REQUEST);
-                log.warn("Can't collect attached files. Trace: {}. Span: {}. Error: {}",
-                        context.getTraceId(), context.getSpanId(), e.getMessage());
-                return;
-            }
-
-            if (deployment instanceof Assistant) {
-                try {
-                    Map.Entry<Buffer, Map<String, String>> enhancedRequest = enhanceAssistantRequest(context, tree);
-                    context.setRequestBody(enhancedRequest.getKey());
-                    context.setRequestHeaders(enhancedRequest.getValue());
-                } catch (HttpException e) {
-                    respond(e.getStatus(), e.getMessage());
-                    log.warn("Can't enhance assistant request. Trace: {}. Span: {}. Error: {}",
-                            context.getTraceId(), context.getSpanId(), e.getMessage());
-                    return;
-                } catch (Throwable e) {
-                    respond(HttpStatus.BAD_REQUEST);
-                    log.warn("Can't enhance assistant request. Trace: {}. Span: {}. Error: {}",
-                            context.getTraceId(), context.getSpanId(), e.getMessage());
-                    return;
-                }
-            }
-            if (deployment instanceof Model) {
-                try {
-                    context.setRequestBody(enhanceModelRequest(context, tree));
-                } catch (Throwable e) {
-                    context.respond(HttpStatus.BAD_REQUEST);
-                    log.warn("Can't enhance model request. Trace: {}. Span: {}. Error: {}",
-                            context.getTraceId(), context.getSpanId(), e.getMessage());
-                    return;
-                }
             }
         } catch (IOException e) {
             respond(HttpStatus.BAD_REQUEST);
@@ -258,26 +224,6 @@ public class DeploymentPostController {
         }
 
         sendRequest();
-    }
-
-    private void processAttachedFile(String url) {
-        ResourceDescription resource = getResourceDescription(url);
-        if (resource == null) {
-            return;
-        }
-        String resourceUrl = resource.getUrl();
-        ApiKeyData sourceApiKeyData = context.getApiKeyData();
-        ApiKeyData destApiKeyData = context.getProxyApiKeyData();
-        AccessService accessService = proxy.getAccessService();
-        if (sourceApiKeyData.getAttachedFiles().contains(resourceUrl) || accessService.hasReadAccess(resource, context)) {
-            if (resource.isFolder()) {
-                destApiKeyData.getAttachedFolders().add(resourceUrl);
-            } else {
-                destApiKeyData.getAttachedFiles().add(resourceUrl);
-            }
-        } else {
-            throw new HttpException(HttpStatus.FORBIDDEN, "Access denied to the file %s".formatted(url));
-        }
     }
 
     @SneakyThrows
@@ -512,98 +458,6 @@ public class DeploymentPostController {
         String endpoint = deployment.getEndpoint();
         String query = request.query();
         return endpoint + (query == null ? "" : "?" + query);
-    }
-
-    private static Map.Entry<Buffer, Map<String, String>> enhanceAssistantRequest(ProxyContext context, ObjectNode tree)
-            throws Exception {
-        Config config = context.getConfig();
-        Assistant assistant = (Assistant) context.getDeployment();
-
-
-        ArrayNode messages = (ArrayNode) tree.get("messages");
-        if (assistant.getPrompt() != null) {
-            deletePrompt(messages);
-            insertPrompt(messages, assistant.getPrompt());
-        }
-
-        Set<String> names = new LinkedHashSet<>(assistant.getAddons());
-        ArrayNode addons = (ArrayNode) tree.get("addons");
-
-        if (addons == null) {
-            addons = tree.putArray("addons");
-        }
-
-        for (JsonNode addon : addons) {
-            String name = addon.get("name").asText("");
-            names.add(name);
-        }
-
-        addons.removeAll();
-        Map<String, String> headers = new HashMap<>();
-        int addonIndex = 0;
-        for (String name : names) {
-            Addon addon = config.getAddons().get(name);
-            if (addon == null) {
-                throw new HttpException(HttpStatus.NOT_FOUND, "No addon: " + name);
-            }
-
-            if (!DeploymentController.hasAccess(context, addon)) {
-                throw new HttpException(HttpStatus.FORBIDDEN, "Forbidden addon: " + name);
-            }
-
-            String url = addon.getEndpoint();
-            addons.addObject().put("url", url).put("name", name);
-            if (addon.getToken() != null && !addon.getToken().isBlank()) {
-                headers.put("x-addon-token-" + addonIndex, addon.getToken());
-            }
-            ++addonIndex;
-        }
-
-        String name = tree.get("model").asText(null);
-        Model model = config.getModels().get(name);
-
-        if (model == null) {
-            throw new HttpException(HttpStatus.NOT_FOUND, "No model: " + name);
-        }
-
-        if (!DeploymentController.hasAccess(context, model)) {
-            throw new HttpException(HttpStatus.FORBIDDEN, "Forbidden model: " + name);
-        }
-
-        Buffer updatedBody = Buffer.buffer(ProxyUtil.MAPPER.writeValueAsBytes(tree));
-        return Map.entry(updatedBody, headers);
-    }
-
-    private static Buffer enhanceModelRequest(ProxyContext context, ObjectNode tree) throws Exception {
-        Model model = (Model) context.getDeployment();
-        String overrideName = model.getOverrideName();
-        Buffer requestBody = context.getRequestBody();
-
-        if (overrideName == null) {
-            return requestBody;
-        }
-
-        tree.remove("model");
-        tree.put("model", overrideName);
-
-        return Buffer.buffer(ProxyUtil.MAPPER.writeValueAsBytes(tree));
-    }
-
-    private static ObjectNode insertPrompt(ArrayNode messages, String prompt) {
-        return messages.insertObject(0)
-                .put("role", "system")
-                .put("content", prompt);
-    }
-
-    private static void deletePrompt(ArrayNode messages) {
-        for (int i = messages.size() - 1; i >= 0; i--) {
-            JsonNode message = messages.get(i);
-            String role = message.get("role").asText("");
-
-            if ("system".equals(role)) {
-                messages.remove(i);
-            }
-        }
     }
 
     private static boolean isBaseAssistant(Deployment deployment) {
