+++ src/main/java/com/epam/aidial/core/data/DeploymentData.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.Data;
 
 import java.util.List;
+import java.util.Map;
 
 @Data
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -29,4 +30,5 @@ public class DeploymentData {
     private FeaturesData features = new FeaturesData();
     private List<String> inputAttachmentTypes;
     private Integer maxInputAttachments;
+    private Map<String, String> defaults;
 }
