+++ src/main/java/com/epam/aidial/core/function/BaseFunction.java
@@ -1,0 +1,16 @@
+package com.epam.aidial.core.function;
+
+import com.epam.aidial.core.Proxy;
+import com.epam.aidial.core.ProxyContext;
+
+import java.util.function.Function;
+
+public abstract class BaseFunction<T> implements Function<T, Throwable> {
+    protected final Proxy proxy;
+    protected final ProxyContext context;
+
+    public BaseFunction(Proxy proxy, ProxyContext context) {
+        this.proxy = proxy;
+        this.context = context;
+    }
+}
